change project struture and refactory all code
