+++ .env.example
@@ -1,0 +1,2 @@
+prefix = v
+token = discord-bot-token
