--- commands/speak.js
@@ -1,3 +1,0 @@
-exports.run = (client, msg, args) => {
-    msg.channel.send('kk eae men')
-}
