--- events/ready.js
@@ -1,3 +1,0 @@
-exports.run = client => {
-    console.log('Hello guys!')
-}
