+++ source/commandResources/file/archetypes/basic.vue
