+++ source/commands/speak.js
@@ -1,0 +1,3 @@
+exports.run = (client, channel, args) => {
+    channel.send('kk eae men')
+}
