+++ source/events/ready.js
@@ -1,0 +1,3 @@
+exports.run = client => {
+    console.log('Vuebot running...')
+}
