+++ source/index.js
@@ -1,0 +1,24 @@
+/* trickstival - 11-3-2017 */
+
+require('dotenv').config()
+
+const Discord = require('discord.js')
+const clientBot = new Discord.Client()
+const fs = require('fs')
+
+/**
+ * Get all files from events folder and register each event in discover bot with name equal filename
+ */
+fs.readdir('./events/', (err, files) => {
+    if(err)
+        throw err
+
+    files.forEach(file => {
+        const eventName = file.replace('.js', '')
+        const eventFunction = require(`./events/${file}`)
+        
+        clientBot.on(eventName, (...args) => eventFunction.run(clientBot, ...args))
+    })
+})
+
+clientBot.login(process.env.token)
