--- tmp/file.vue
@@ -1,1 +1,0 @@
-test6
